organize directive to remove iOS references
