+++ CutTheRope/commons/CtrRenderer.cs
@@ -4,11 +4,11 @@ using System.Diagnostics;
 
 using CutTheRope.desktop;
 using CutTheRope.game;
+using CutTheRope.Helpers;
 using CutTheRope.iframework;
 using CutTheRope.iframework.core;
 using CutTheRope.iframework.platform;
 using CutTheRope.iframework.visual;
-using CutTheRope.ios;
 
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Input.Touch;
