+++ CutTheRope/game/BoxOpenClose.cs
@@ -1,11 +1,11 @@
 using System;
 using System.Globalization;
 
+using CutTheRope.Helpers;
 using CutTheRope.iframework;
 using CutTheRope.iframework.core;
 using CutTheRope.iframework.helpers;
 using CutTheRope.iframework.visual;
-using CutTheRope.ios;
 
 namespace CutTheRope.game
 {
