+++ CutTheRope/game/MenuController.cs
@@ -5,12 +5,12 @@ using System.Reflection;
 
 using CutTheRope.commons;
 using CutTheRope.desktop;
+using CutTheRope.Helpers;
 using CutTheRope.iframework;
 using CutTheRope.iframework.core;
 using CutTheRope.iframework.helpers;
 using CutTheRope.iframework.media;
 using CutTheRope.iframework.visual;
-using CutTheRope.ios;
 
 using Microsoft.Xna.Framework.Input;
 using Microsoft.Xna.Framework.Input.Touch;
