+++ CutTheRope/iframework/helpers/Mover.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
 
+using CutTheRope.Helpers;
 using CutTheRope.iframework.core;
-using CutTheRope.ios;
 
 namespace CutTheRope.iframework.helpers
 {
