+++ CutTheRope/iframework/visual/Font.cs
@@ -1,6 +1,6 @@
 using System;
 
-using CutTheRope.ios;
+using CutTheRope.Helpers;
 
 namespace CutTheRope.iframework.visual
 {
